+++ src/provider/kubernetes/gateway.rs
@@ -1093,6 +1093,7 @@ fn rule_to_entrypoints(
                     methods: Vec::new(),
                     acme: None,
                     plugins: Vec::new(),
+                    error_pages: std::collections::BTreeMap::new(),
                 },
                 source: Some(id),
             }